Rewrite LayoutComponents without constructors
