+++ packages/bento-design-system/src/index.ts
@@ -30,7 +30,12 @@ export * from "./Icons/svgIconProps";
 export * from "./Illustrations";
 export * from "./Illustrations/svgIllustrationProps";
 export * from "./InlineLoader/InlineLoader";
-export * from "./Layout/createLayoutComponents";
+export * from "./Layout/Bleed";
+export * from "./Layout/Columns";
+export * from "./Layout/Inline";
+export * from "./Layout/Inset";
+export * from "./Layout/Stack";
+export * from "./Layout/Tiles";
 export * from "./Link/Link";
 export * from "./List/createListComponents";
 export * from "./Menu/createMenu";
